# docs/swagger.yaml
basePath: /api/v1
definitions:
  types.CommentResponse:
    properties:
      comment_id:
        type: integer
      content_text:
        type: string
      post_id:
        type: integer
      user_id:
        type: integer
    type: object
  types.CreatePostCommentRequest:
    properties:
      content_text:
        type: string
    required:
    - content_text
    type: object
  types.CreatePostRequest:
    properties:
      content_image_path:
        items:
          type: string
        type: array
      content_text:
        type: string
      visible:
        type: boolean
    required:
    - content_text
    type: object
...
  types.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  types.PostDetailInfoResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/types.CommentResponse'
        type: array
      content_image_path:
        items:
          type: string
        type: array
      content_text:
        type: string
      created_at:
        type: string
      post_id:
        type: integer
      user_id:
        type: integer
      users_liked:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact:
    email: maxuanquang@gmail.com
    name: Quang Ma
    url: https://www.linkedin.com/in/maxuanquang
  description: A simple social network management service API in Go using Gin framework
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://maxuanquang.dev/gin-social-network-tos
  title: Gin Social Network Service
  version: "1.0"
paths:
...
  /posts/:
    delete:
      consumes:
      - application/json
      description: delete post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: delete post
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: create new post
      parameters:
      - description: Create post parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: create new post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: edit post information
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: edit post information
      tags:
      - posts
  /posts/{post_id}:
    get:
      consumes:
      - application/json
      description: get post detail information
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PostDetailInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: get post detail information
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: comment to post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      - description: Comment's content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.CreatePostCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PostDetailInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: comment to post
      tags:
      - posts
  /posts/{post_id}/likes:
    post:
      consumes:
      - application/json
      description: like post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: like post
      tags:
      - posts
...
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      admin: "\t\t\t\t\t\t\tGrants read and write access to administrative information"
      read: "\t\t\t\t\t\t\t\tGrants read access"
      write: "\t\t\t\t\t\t\tGrants write access"
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"